# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MLOVI - Multidisciplinar: BACKEND API V1",
        "description": "Description of your API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Main (production) server"
        }
    ],
    "paths": {
...
        "/responsaveis": {
            "get": {
                "tags": [
                    "Responsáveis"
                ],
                "summary": "Listar todos os responsáveis",
                "operationId": "getResponsavelList",
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida"
                    },
                    "401": {
                        "description": "Não Autenticado"
                    },
                    "403": {
                        "description": "Proibido"
                    }
                }
            },
            "post": {
                "tags": [
                    "Responsáveis"
                ],
                "summary": "Cadastrar um novo responsavel",
                "description": "Cria um novo registro de responsavel com os dados fornecidos.",
                "operationId": "storeResponsavel",
                "requestBody": {
                    "description": "Dados do responsavel",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "cpf",
                                    "email",
                                    "data_nascimento",
                                    "sexo"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "João da Silva"
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "example": "123.456.789-11"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "joao123@example.com"
                                    },
                                    "data_nascimento": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2000-01-01"
                                    },
                                    "sexo": {
                                        "type": "string",
                                        "example": "M"
                                    },
                                    "foto": {
                                        "type": "string",
                                        "example": "arquivo.jpg",
                                        "nullable": true
                                    },
                                    "cep": {
                                        "type": "string",
                                        "example": "00000-000",
                                        "nullable": true
                                    },
                                    "endereco": {
                                        "type": "string",
                                        "example": "Rua Exemplo",
                                        "nullable": true
                                    },
                                    "numero": {
                                        "type": "string",
                                        "example": "123",
                                        "nullable": true
                                    },
                                    "complemento": {
                                        "type": "string",
                                        "example": "Apto 101",
                                        "nullable": true
                                    },
                                    "bairro": {
                                        "type": "string",
                                        "example": "Centro",
                                        "nullable": true
                                    },
                                    "cidade": {
                                        "type": "string",
                                        "example": "Cidade Exemplo",
                                        "nullable": true
                                    },
                                    "estado": {
                                        "type": "string",
                                        "example": "Estado Exemplo",
                                        "nullable": true
                                    },
                                    "paciente_id": {
                                        "type": "intenger",
                                        "example": "1",
                                        "nullable": true
                                    },
                                    "parentesco": {
                                        "type": "string",
                                        "example": "Pai",
                                        "nullable": true
                                    },
                                    "relacionamento": {
                                        "type": "string",
                                        "example": "Responsavel",
                                        "nullable": true
                                    },
                                    "responsavelFinanceiro": {
                                        "type": "boolean",
                                        "example": "true"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Responsavel criado com sucesso"
                    },
                    "400": {
                        "description": "Erro na requisição"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Proibido"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/responsaveis/{id}": {
            "get": {
                "tags": [
                    "Responsáveis"
                ],
                "summary": "Buscar responsavel por ID",
                "operationId": "getResponsavelById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do responsavel",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida"
                    },
                    "401": {
                        "description": "Não Autenticado"
                    },
                    "403": {
                        "description": "Proibido"
                    },
                    "404": {
                        "description": "Responsavel não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Responsáveis"
                ],
                "summary": "Atualizar responsavel",
                "operationId": "updateResponsavel",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do responsavel",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Dados do responsavel a serem alterados. Verifique os campos no Schema",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "José Silva"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida"
                    },
                    "401": {
                        "description": "Não Autenticado"
                    },
                    "403": {
                        "description": "Proibido"
                    },
                    "404": {
                        "description": "Responsavel não encontrado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Responsáveis"
                ],
                "summary": "Excluir responsavel",
                "operationId": "deleteResponsavel",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do responsavel",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Responsavel excluído com sucesso"
                    },
                    "401": {
                        "description": "Não Autenticado"
                    },
                    "403": {
                        "description": "Proibido"
                    },
                    "404": {
                        "description": "Responsavel não encontrado"
                    }
                }
            }
        },
...
    },
...
}